+++ doc/changelog.rst
@@ -1,6 +1,14 @@
 Changelog
 =========
 
+Changes in Version 0.7.0
+------------------------
+
+- Added optional support for passing in Channel Binding Tokens (RFC 5929) into
+  :func:`winkerberos.authGSSClientStep`. The binding token structure can be
+  built using :func:`winkerberos.channelBindings` (see the example
+  for more details).
+
 Changes in Version 0.6.0
 ------------------------